[Refactor]: `챌린지 그룹`에 대한 순환의존성 제거
