+++ zzansuni-api-server/app/src/main/java/org/haedal/zzansuni/domain/challengegroup/port/ChallengeGroupReader.java
@@ -3,7 +3,7 @@ package org.haedal.zzansuni.domain.challengegroup.port;
 import org.haedal.zzansuni.domain.challengegroup.ChallengeCategory;
 import org.haedal.zzansuni.domain.challengegroup.ChallengeGroup;
 import org.haedal.zzansuni.domain.challengegroup.application.ChallengeGroupModel;
-import org.haedal.zzansuni.domain.userchallenge.ChallengeGroupUserExp;
+import org.haedal.zzansuni.domain.challengegroup.ChallengeGroupUserExp;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
