+++ zzansuni-api-server/app/src/main/java/org/haedal/zzansuni/domain/userchallenge/ChallengeStatus.java
@@ -1,4 +1,4 @@
-package org.haedal.zzansuni.domain.challengegroup;
+package org.haedal.zzansuni.domain.userchallenge;
 
 import lombok.RequiredArgsConstructor;
