+++ zzansuni-api-server/app/src/main/java/org/haedal/zzansuni/domain/userchallenge/application/AddUserExpByVerificationUseCase.java
@@ -3,7 +3,7 @@ package org.haedal.zzansuni.domain.userchallenge.application;
 import lombok.RequiredArgsConstructor;
 import org.haedal.zzansuni.domain.challengegroup.port.ChallengeGroupReader;
 import org.haedal.zzansuni.domain.userchallenge.AddUserExpByVerificationEvent;
-import org.haedal.zzansuni.domain.userchallenge.ChallengeGroupUserExp;
+import org.haedal.zzansuni.domain.challengegroup.ChallengeGroupUserExp;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
