+++ zzansuni-api-server/app/src/main/java/org/haedal/zzansuni/domain/userchallenge/application/UserChallengeService.java
@@ -8,8 +8,8 @@ import org.haedal.zzansuni.domain.userchallenge.AddUserExpByVerificationEvent;
 import org.haedal.zzansuni.domain.userchallenge.UserChallenge;
 import org.haedal.zzansuni.domain.userchallenge.port.UserChallengeReader;
 import org.haedal.zzansuni.domain.userchallenge.port.UserChallengeStore;
-import org.haedal.zzansuni.domain.userchallenge.ChallengeGroupUserExp;
-import org.haedal.zzansuni.domain.userchallenge.port.ChallengeGroupUserExpStore;
+import org.haedal.zzansuni.domain.challengegroup.ChallengeGroupUserExp;
+import org.haedal.zzansuni.domain.challengegroup.port.ChallengeGroupUserExpStore;
 import org.haedal.zzansuni.domain.challengegroup.Challenge;
 import org.haedal.zzansuni.domain.challengegroup.ChallengeCommand;
 import org.haedal.zzansuni.domain.challengegroup.application.ChallengeModel;
