+++ zzansuni-api-server/app/src/main/java/org/haedal/zzansuni/infrastructure/challengegroup/ChallengeGroupUserExpRepository.java
@@ -1,6 +1,6 @@
-package org.haedal.zzansuni.infrastructure.userchallenge;
+package org.haedal.zzansuni.infrastructure.challengegroup;
 
-import org.haedal.zzansuni.domain.userchallenge.ChallengeGroupUserExp;
+import org.haedal.zzansuni.domain.challengegroup.ChallengeGroupUserExp;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ChallengeGroupUserExpRepository extends JpaRepository<ChallengeGroupUserExp, Long> {
