+++ zzansuni-api-server/app/src/test/java/org/haedal/zzansuni/domain/challengegroup/challenge/ChallengeRecordServiceTest.java
@@ -3,21 +3,18 @@ package org.haedal.zzansuni.domain.challengegroup.challenge;
 import java.time.LocalDate;
 
 import org.haedal.zzansuni.domain.challengegroup.*;
+import org.haedal.zzansuni.domain.userchallenge.*;
 import org.haedal.zzansuni.domain.userchallenge.application.ChallengeRecordService;
 import org.haedal.zzansuni.domain.challengegroup.application.ChallengeModel.ChallengeRecord;
 import org.haedal.zzansuni.domain.challengegroup.port.ChallengeReader;
-import org.haedal.zzansuni.domain.userchallenge.ChallengeReview;
 import org.haedal.zzansuni.domain.userchallenge.application.ChallengeReviewModel;
 import org.haedal.zzansuni.domain.userchallenge.application.ChallengeReviewService;
 import org.haedal.zzansuni.domain.userchallenge.port.ChallengeReviewReader;
 import org.haedal.zzansuni.domain.userchallenge.port.ChallengeReviewStore;
-import org.haedal.zzansuni.domain.userchallenge.UserChallenge;
 import org.haedal.zzansuni.domain.userchallenge.port.UserChallengeReader;
-import org.haedal.zzansuni.domain.userchallenge.ChallengeVerification;
 import org.haedal.zzansuni.domain.userchallenge.application.ChallengeVerificationModel;
 import org.haedal.zzansuni.domain.userchallenge.port.ChallengeVerificationReader;
 import org.haedal.zzansuni.domain.user.User;
-import org.haedal.zzansuni.domain.userchallenge.ChallengeVerificationStatus;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
