+++ zzansuni-api-server/app/src/test/java/org/haedal/zzansuni/domain/challengegroup/userchallenge/UserChallengeRecordServiceTest.java
@@ -7,13 +7,14 @@ import org.haedal.zzansuni.domain.challengegroup.application.ChallengeModel;
 import org.haedal.zzansuni.domain.challengegroup.port.ChallengeGroupReader;
 import org.haedal.zzansuni.domain.challengegroup.port.ChallengeReader;
 import org.haedal.zzansuni.domain.userchallenge.application.AddUserExpByVerificationUseCase;
+import org.haedal.zzansuni.domain.userchallenge.ChallengeStatus;
 import org.haedal.zzansuni.domain.userchallenge.port.ChallengeReviewReader;
 import org.haedal.zzansuni.domain.userchallenge.UserChallenge;
 import org.haedal.zzansuni.domain.userchallenge.port.UserChallengeReader;
 import org.haedal.zzansuni.domain.userchallenge.application.UserChallengeService;
 import org.haedal.zzansuni.domain.userchallenge.port.UserChallengeStore;
-import org.haedal.zzansuni.domain.userchallenge.port.ChallengeGroupUserExpStore;
-import org.haedal.zzansuni.domain.userchallenge.ChallengeGroupUserExp;
+import org.haedal.zzansuni.domain.challengegroup.port.ChallengeGroupUserExpStore;
+import org.haedal.zzansuni.domain.challengegroup.ChallengeGroupUserExp;
 import org.haedal.zzansuni.domain.user.User;
 import org.haedal.zzansuni.domain.user.UserReader;
 import org.haedal.zzansuni.global.security.Role;
